pyproject.toml
[project]
name = "forgestack"
version = "0.1.0"
description = "ForgeStack modular application generator"
requires-python = ">=3.10"

dependencies = [
    "click",
    "pyyaml",
    "jinja2",
    "networkx",
    "rich"
]

[project.scripts]
devmake = "forgestack.cli.devmake:main"